feat: refactor code.
