+++ Dockerfile
@@ -1,7 +1,6 @@
-FROM python:3.12
+FROM python:3.11
 
-RUN mkdir /data
-ENV CYCAX_VAR_DIR=/data
+RUN apt-get update; apt-get install -y libxxf86vm-dev libxfixes-dev libxi-dev libxkbcommon-dev libgl-dev
 RUN mkdir /app
 WORKDIR /app
 COPY . .
