+++ src/cycax_blender_worker/config.py
@@ -1,0 +1,9 @@
+from pathlib import Path
+
+from pydantic_settings import BaseSettings, SettingsConfigDict
+
+
+class Config(BaseSettings):
+    model_config = SettingsConfigDict(env_prefix="cycax_")
+    server: str
+    temp_dir: Path = Path("/tmp/cycax_blender_worker")  # noqa: S108
